+++ tests/unit/charts/test_logprep_chart.py
@@ -4,72 +4,25 @@
 
 import subprocess
 from tempfile import NamedTemporaryFile
-from typing import Dict, List, Optional
+from typing import Dict, Optional
 
-from attrs import define, field, validators
-
-from logprep.filter.expression.filter_expression import KeyDoesNotExistError
-from logprep.filter.lucene_filter import LuceneFilter
 from logprep.util.configuration import yaml
-from logprep.util.helper import get_dotted_field_value
+from logprep.util.event import Documents
 
 LOGPREP_CHART_PATH = "charts/logprep"
-
-
-def convert_manifests(data: bytes | list) -> List[Dict]:
-    if isinstance(data, bytes):
-        data = yaml.load_all(data)
-    return [
-        Manifest(manifest_dict) if isinstance(manifest_dict, dict) else manifest_dict
-        for manifest_dict in data
-        if manifest_dict
-    ]
-
-
-@define
-class Manifests:
-
-    _manifests: List["Manifest"] = field(
-        validator=validators.instance_of(list), converter=convert_manifests
-    )
-
-    def by_query(self, query: str) -> "Manifests":
-        return Manifests([manifest for manifest in self._manifests if manifest.query(query)])
-
-    def __len__(self):
-        return len(self._manifests)
-
-
-@define
-class Manifest:
-
-    _manifest: Dict = field(validator=validators.instance_of(dict))
-
-    def query(self, query: str) -> bool:
-        filter_expression = LuceneFilter.create(query)
-        try:
-            return filter_expression.does_match(self._manifest)
-        except KeyDoesNotExistError:
-            return False
-
-    def __getitem__(self, key):
-        return get_dotted_field_value(self._manifest, key)
-
-    def __contains__(self, key):
-        return self.query(key)
 
 
 class TestBaseChartTest:
 
     @staticmethod
-    def render_chart(release, values: Optional[Dict] = None) -> Manifests:
+    def render_chart(release, values: Optional[Dict] = None) -> Documents:
         """Render a Helm chart with the given values and return the Kubernetes objects."""
         values = values or {}
         with NamedTemporaryFile() as tmp_file:
             content = yaml.dump(values)
             tmp_file.write(content.encode())
             tmp_file.flush()
-            return Manifests(
+            return Documents(
                 subprocess.check_output(
                     ["helm", "template", release, LOGPREP_CHART_PATH, "--values", tmp_file.name]
                 )
@@ -82,9 +35,9 @@ class TestDefaultValues(TestBaseChartTest):
         self.manifests = self.render_chart("logprep")
 
     def test_labels_are_set(self):
-        for manifest in self.manifests._manifests:
+        for manifest in self.manifests:
             assert "metadata.labels" in manifest
 
     def test_application_label_is_set(self):
-        for manifest in self.manifests._manifests:
+        for manifest in self.manifests:
             assert manifest["metadata.labels"]["app.kubernetes.io/name"] == "logprep-logprep"
